Fix Xen address translation.
It was a bad idea to use cr3 from pcpu 0 to translate Xen virtual addresses.
Instead we should use value of idle_pg_table symbol to get to machine
address of the head of page directory. That value must be used for Xen adress
translation, except directly mapped addresses of xen text and xen 1:1 virt to
machine area.